requirements.txt
numpy>=1.24
pandas>=2.0
yfinance>=0.2.28
alpaca-py>=0.21
scipy>=1.10
matplotlib>=3.7
scikit-learn>=1.2
tqdm>=4.65
flask>=3.0
flask-socketio>=5.3
eventlet>=0.35
requests>=2.28